move kr writer test to the correct file
